requirements.txt
transformers
requests
pillow
torch
customtkinter